refactor: 로그인이 되어있지 않을 경우 수행할 로직을 유틸함수로 생성
